None: Unify namespace and remove compiler warnings

See merge request TensorRT/tensorrt-edge-llm/tensorrt-edge-llm!230

+++ cpp/common/tensor.cpp
@@ -1,3 +1,15 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: LicenseRef-NvidiaProprietary
+ *
+ * NVIDIA CORPORATION, its affiliates and licensors retain all intellectual
+ * property and proprietary rights in and to this material, related
+ * documentation and any modifications thereto. Any use, reproduction,
+ * disclosure or distribution of this material and related documentation
+ * without an express license agreement from NVIDIA CORPORATION or
+ * its affiliates is strictly prohibited.
+ */
+
 #include "tensor.h"
 
 #include "common.h"
@@ -230,7 +242,7 @@ bool Tensor::reshape(Coords shape) noexcept
         return false;
     }
 
-    if (shape.volume() * getTypeSize(mDataType) > memoryCapacity)
+    if (static_cast<int64_t>(shape.volume() * getTypeSize(mDataType)) > memoryCapacity)
     {
         return false;
     }
